requirements.txt
fastapi
uvicorn
python-dotenv
motor
pydantic
google-generativeai
openai
